Fog rangeModifier

+++ Tower.java
@@ -3,7 +3,7 @@ import java.util.ArrayList;
 public class Tower extends Construct {
 	private int damage;
 	private int fireRate;
-	private int rangeModifier;
+	private double rangeModifier;
 	private int range;
 	private FieldTile towerLocation;
 
@@ -81,7 +81,7 @@ public class Tower extends Construct {
 	 * Beállítja a torony hatótávolságának módosítóját
 	 * @param rangeModifier a torony hatótávának módosítója
 	 */
-	public void setRangeModifier(int rangeModifier) {
+	public void setRangeModifier(double rangeModifier) {
 		this.rangeModifier = rangeModifier;
 	}
